extract pup-owned helpers out of the generator-owned client.rs/util.rs
openapi-transformer's pup generator copies client.rs/util.rs byte-for-byte
from a bundled snapshot on every regeneration, but only make_api!/
make_api_no_auth! and util::read_json_file are actually generator-relevant.
Everything else was pup's own hand-written command logic that just happened
to live in those files, so a routine feature PR (#654, event posting) got
silently reverted the next time the client was regenerated (pup#651).

Split both files along the generator/hand-written boundary, mirroring how
main.rs/generated.rs already separate hand-written and generated code:

- client.rs keeps only what make_api!/make_api_no_auth! need:
  BearerAuthMiddleware/UserAgentMiddleware, make_dd_config/make_dd_client,
  the two macros, UNSTABLE_OPS.
- New raw_client.rs holds the rest: HttpError/HttpResponse, AuthType/
  get_auth_type, the OAuth-exclusion table + apply_auth, parse_response_json,
  and every raw_* HTTP helper used by ~16 hand-written commands for
  endpoints not covered by the typed SDK.
- util.rs keeps only read_json_file. New util_ext.rs holds time/duration
  parsing, the monitor-diff helpers, parse_compute_raw, and percent_encode.

Pure move: identical function bodies/signatures, only import paths changed.
