+++ src/commands/app_builder.rs
@@ -8,6 +8,7 @@ use datadog_api_client::datadogV2::model::{
 use crate::config::Config;
 use crate::formatter;
 use crate::util;
+use crate::util_ext;
 
 pub async fn list(cfg: &Config, query: Option<&str>) -> Result<()> {
     let api = crate::make_api!(AppBuilderAPI, cfg);
@@ -24,7 +25,7 @@ pub async fn list(cfg: &Config, query: Option<&str>) -> Result<()> {
 
 pub async fn get(cfg: &Config, app_id: &str) -> Result<()> {
     let api = crate::make_api!(AppBuilderAPI, cfg);
-    let uuid = util::parse_uuid(app_id, "app")?;
+    let uuid = util_ext::parse_uuid(app_id, "app")?;
     let resp = api
         .get_app(uuid, Default::default())
         .await
@@ -44,7 +45,7 @@ pub async fn create(cfg: &Config, file: &str) -> Result<()> {
 
 pub async fn update(cfg: &Config, app_id: &str, file: &str) -> Result<()> {
     let api = crate::make_api!(AppBuilderAPI, cfg);
-    let uuid = util::parse_uuid(app_id, "app")?;
+    let uuid = util_ext::parse_uuid(app_id, "app")?;
     let body: UpdateAppRequest = util::read_json_file(file)?;
     let resp = api
         .update_app(uuid, body)
@@ -55,7 +56,7 @@ pub async fn update(cfg: &Config, app_id: &str, file: &str) -> Result<()> {
 
 pub async fn delete(cfg: &Config, app_id: &str) -> Result<()> {
     let api = crate::make_api!(AppBuilderAPI, cfg);
-    let uuid = util::parse_uuid(app_id, "app")?;
+    let uuid = util_ext::parse_uuid(app_id, "app")?;
     api.delete_app(uuid)
         .await
         .map_err(|e| anyhow::anyhow!("failed to delete app: {e:?}"))?;
@@ -68,7 +69,7 @@ pub async fn delete_batch(cfg: &Config, app_ids: &[String]) -> Result<()> {
     let items: Result<Vec<_>> = app_ids
         .iter()
         .map(|id| {
-            let uuid = util::parse_uuid(id, "app")?;
+            let uuid = util_ext::parse_uuid(id, "app")?;
             Ok(DeleteAppsRequestDataItems::new(
                 uuid,
                 AppDefinitionType::APPDEFINITIONS,
@@ -85,7 +86,7 @@ pub async fn delete_batch(cfg: &Config, app_ids: &[String]) -> Result<()> {
 
 pub async fn publish(cfg: &Config, app_id: &str) -> Result<()> {
     let api = crate::make_api!(AppBuilderAPI, cfg);
-    let uuid = util::parse_uuid(app_id, "app")?;
+    let uuid = util_ext::parse_uuid(app_id, "app")?;
     let resp = api
         .publish_app(uuid)
         .await
@@ -95,7 +96,7 @@ pub async fn publish(cfg: &Config, app_id: &str) -> Result<()> {
 
 pub async fn unpublish(cfg: &Config, app_id: &str) -> Result<()> {
     let api = crate::make_api!(AppBuilderAPI, cfg);
-    let uuid = util::parse_uuid(app_id, "app")?;
+    let uuid = util_ext::parse_uuid(app_id, "app")?;
     api.unpublish_app(uuid)
         .await
         .map_err(|e| anyhow::anyhow!("failed to unpublish app: {e:?}"))?;
