+++ src/commands/workflows.rs
@@ -9,6 +9,7 @@ use datadog_api_client::datadogV2::api_workflow_automation::{
 use crate::config::Config;
 use crate::formatter::{self, Metadata};
 use crate::util;
+use crate::util_ext;
 
 // ---------------------------------------------------------------------------
 // Helper: build a WorkflowAutomationAPI
@@ -119,7 +120,7 @@ pub async fn run(
     eprintln!("Instance {instance_id} started, waiting for completion...");
 
     let timeout_duration =
-        std::time::Duration::from_millis(crate::util::parse_duration_to_millis(timeout)? as u64);
+        std::time::Duration::from_millis(util_ext::parse_duration_to_millis(timeout)? as u64);
     let start = std::time::Instant::now();
 
     loop {
